+++ src/main/java/com/telepathicgrunt/the_bumblezone/features/BzFeatures.java
@@ -3,10 +3,7 @@ package com.telepathicgrunt.the_bumblezone.features;
 import com.telepathicgrunt.the_bumblezone.Bumblezone;
 import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.registry.Registry;
-import net.minecraft.util.registry.WorldGenRegistries;
-import net.minecraft.world.gen.feature.ConfiguredFeature;
 import net.minecraft.world.gen.feature.Feature;
-import net.minecraft.world.gen.feature.IFeatureConfig;
 import net.minecraft.world.gen.feature.NoFeatureConfig;
 
 public class BzFeatures {